requirements.txt
asgiref==3.3.4
dj-database-url==0.5.0
Django==3.1.7
django-cors-headers==3.7.0
django-filter==2.4.0
django-params-validator==1.1.10
django-storages==1.11.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
gunicorn==20.1.0
Pillow==8.1.2
PyJWT==2.0.1
pytz==2021.1
sqlparse==0.4.1
whitenoise==5.2.0
